@@ -9,8 +9,9 @@
     "build": "rm -rf ./dist && npm run compile",
     "generate": "graphql-codegen",
     "dev": "ts-node-dev --respawn ./src/index.ts",
-    "test": "jest -c ./jest.config.js",
+    "test": "jest --coverage",
     "test-tsc": "jest -c ./jest.tsc.config.js",
+    "lint": "npx eslint .",
     "start": "npm run build && nodemon ./dist/index.js",
     "aws_start": "node ./dist/index.js"
   },
@@ -32,7 +33,7 @@
     "@graphql-tools/merge": "^9.0.3",
     "@graphql-tools/mock": "^9.0.0",
     "@graphql-tools/schema": "^10.0.0",
-    "@types/express": "^4.17.21",
+    "@node-oauth/express-oauth-server": "^4.0.0",
     "aws-sdk": "^2.1598.0",
     "bcryptjs": "^2.4.3",
     "casual": "^1.6.2",
@@ -45,19 +46,29 @@
     "jsonwebtoken": "^9.0.2",
     "mysql2": "^3.9.7",
     "pino": "^9.0.0",
-    "uuid": "^9.0.1"
+    "uuid": "^9.0.1",
+    "uuid-random": "^1.3.2"
   },
   "devDependencies": {
+    "@eslint/js": "^9.6.0",
     "@graphql-codegen/cli": "^4.0.1",
     "@graphql-codegen/typescript": "^4.0.1",
     "@graphql-codegen/typescript-resolvers": "^4.0.1",
-    "@types/assert": "^1.5.10",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/eslint__js": "^8.42.3",
+    "@types/express": "^4.17.21",
+    "@types/express-oauth-server": "^2.0.7",
     "@types/jest": "^29.5.12",
+    "@types/oauth2-server": "^3.0.16",
+    "@types/pino": "^7.0.5",
+    "eslint": "^8.57.0",
+    "globals": "^15.6.0",
     "jest": "^29.7.0",
-    "jest-runner-tsc": "^1.6.0",
+    "jest-mock": "^29.7.0",
     "nodemon": "^3.1.0",
-    "ts-jest": "^29.1.2",
+    "ts-jest": "^29.1.4",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.3.3"
+    "typescript": "^5.5.2",
+    "typescript-eslint": "^7.14.1"
   }
 }
