@@ -1,9 +1,7 @@
 import { expressMiddleware } from '@apollo/server/express4';
-
-import { extractToken } from './auth';
-import { mysqlConfig } from '../config';
 import { DMPHubAPI } from '../datasources/dmphub-api';
-import { MysqlDataSource } from '../datasources/mysqlDB';
+import { MySQLDataSource } from '../datasources/mySQLDataSource';
+import { JWTToken, verifyToken } from '../services/tokenService';
 
 export function attachApolloServer(apolloServer, cache, logger) {
   // expressMiddleware accepts the same arguments:
@@ -11,14 +9,20 @@ export function attachApolloServer(apolloServer, cache, logger) {
   return expressMiddleware(apolloServer, {
     context: async ({ req }) => {
       // Extract the token from the incoming request so we can pass it on to the resolvers
-      const token = extractToken(req);
+      const authHeader: string = req?.headers?.authorization || '';
+      const authHdr: string = authHeader.split(' ')[1] || null;
+      const token: JWTToken = authHeader ? verifyToken(authHdr, logger) : null;
+
+console.log(authHdr);
+console.log(token);
 
       return {
+        token,
         // Pass the logger in so it is available to our resolvers and dataSources
         logger,
         dataSources: {
           dmphubAPIDataSource: await new DMPHubAPI({ cache, token }),
-          sqlDataSource: await new MysqlDataSource({ config: mysqlConfig }),
+          sqlDataSource: await MySQLDataSource.getInstance(),
         },
       }
     },
